+++ app/models/bank_cell.rb
@@ -2,4 +2,6 @@
 class BankCell < ApplicationRecord
   belongs_to :bank
   belongs_to :game_item, optional: true
+
+  scope :empty, -> { where game_item_id: nil }
 end