+++ app/policies/guild_policy.rb
@@ -11,4 +11,8 @@ class GuildPolicy < ApplicationPolicy
   def bank?
     user.has_characters_in_guild?(guild_id: record.id)
   end
+
+  def bank_management?
+    user.any_role?(record.id, 'gm', 'ba')
+  end
 end