+++ app/serializers/bank_serializer.rb
@@ -1,0 +1,4 @@
+class BankSerializer < ActiveModel::Serializer
+  attributes :id, :name, :coins
+  has_many :bank_cells, serializer: BankCellSerializer
+end