+++ config/routes.rb
@@ -12,6 +12,7 @@ Rails.application.routes.draw do
         get :form_values, on: :collection
         get :characters_for_request, on: :member
         post :import_bank, on: :member
+        get :bank, on: :member
       end
       resources :roles, only: %i[index]
       resources :dungeons, only: %i[index]