+++ frontend/packs/guild_bank.jsx
@@ -1,0 +1,4 @@
+import GuildBank from "components_react/guild_bank/guild_bank";
+import WebpackerReact from "webpacker-react";
+
+WebpackerReact.setup({GuildBank});